split the logger into two cpp files a loggerCore which is internal and logger.h which has access to the logger in the helics library.
