+++ src/helics/apps/Tracer.cpp
@@ -21,7 +21,7 @@ SPDX-License-Identifier: BSD-3-Clause
 
 #include "../common/JsonProcessingFunctions.hpp"
 #include "../common/fmt_format.h"
-#include "../common/logger.h"
+#include "../common/loggerCore.hpp"
 #include "PrecHelper.hpp"
 #include "Tracer.hpp"
 #include <thread>
@@ -512,7 +512,7 @@ std::shared_ptr<helicsCLI11App> Tracer::buildArgParserApp ()
     auto clone_group =
       app->add_option_group ("cloning", "Options related to endpoint cloning operations and specifications");
     clone_group->add_option ("--clone", "existing endpoints to clone all packets to and from")
-      ->each ([this] (const std::string &clone) {
+      ->each ([this](const std::string &clone) {
           addDestEndpointClone (clone);
           addSourceEndpointClone (clone);
       })
@@ -523,7 +523,7 @@ std::shared_ptr<helicsCLI11App> Tracer::buildArgParserApp ()
       ->add_option (
         "--sourceclone",
         "existing endpoints to capture generated packets from, this argument may be specified multiple time")
-      ->each ([this] (const std::string &clone) { addSourceEndpointClone (clone); })
+      ->each ([this](const std::string &clone) { addSourceEndpointClone (clone); })
       ->delimiter (',')
       ->ignore_underscore ()
       ->type_size (-1);
@@ -531,7 +531,7 @@ std::shared_ptr<helicsCLI11App> Tracer::buildArgParserApp ()
     clone_group
       ->add_option ("--destclone", "existing endpoints to capture all packets with the specified endpoint as a "
                                    "destination, this argument may be specified multiple time")
-      ->each ([this] (const std::string &clone) { addSourceEndpointClone (clone); })
+      ->each ([this](const std::string &clone) { addSourceEndpointClone (clone); })
       ->delimiter (',')
       ->ignore_underscore ()
       ->type_size (-1);
@@ -542,7 +542,7 @@ std::shared_ptr<helicsCLI11App> Tracer::buildArgParserApp ()
     capture_group
       ->add_option ("--tag,--publication,--pub",
                     "tags(publications) to record, this argument may be specified any number of times")
-      ->each ([this] (const std::string &tag) {
+      ->each ([this](const std::string &tag) {
           auto taglist = stringOps::splitlineQuotes (tag);
           for (const auto &tagname : taglist)
           {
@@ -552,7 +552,7 @@ std::shared_ptr<helicsCLI11App> Tracer::buildArgParserApp ()
       ->type_size (-1);
 
     capture_group->add_option ("--endpoints", "endpoints to capture, this argument may be specified multiple time")
-      ->each ([this] (const std::string &ept) {
+      ->each ([this](const std::string &ept) {
           auto eptlist = stringOps::splitlineQuotes (ept);
           for (const auto &eptname : eptlist)
           {
@@ -564,7 +564,7 @@ std::shared_ptr<helicsCLI11App> Tracer::buildArgParserApp ()
     capture_group
       ->add_option ("--capture", "capture all the publications of a particular federate capture=\"fed1;fed2\"  "
                                  "supports multiple arguments or a comma separated list")
-      ->each ([this] (const std::string &capt) {
+      ->each ([this](const std::string &capt) {
           auto captFeds = stringOps::splitlineQuotes (capt);
           for (auto &captFed : captFeds)
           {
